Typing improvements (#146)
* Fixing various typing issues

* Adding pyright check

* Fixing style

* Adding verbose flag to pyright

* Attempting to fix pyright target

+++ src/benchq/problem_embeddings/qpe.py
@@ -24,7 +24,7 @@ def get_single_factorized_qpe_toffoli_and_qubit_cost(
     eri: np.ndarray,
     rank: int,
     allowable_phase_estimation_error: float = 0.001,
-    bits_precision_coefficients: float = 10,
+    bits_precision_coefficients: int = 10,
 ) -> Tuple[int, int]:
     """Get the number of Toffoli gates and logical qubits for single factorized QPE as
     described in PRX Quantum 2, 030305.
